+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,14 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2022-08-26
+
+### Added
+
+- [List] Added `reset` method to the `useRecoilList` hook.
+- [Dialog] Added `reset` method to the `useRecoilDialog` hook.
+- [Filters] Added `reset` method to the `useRecoilFilters` hook.
+- [Pagination] Added `reset` method to the `useRecoilPagination` hook.
+
 ## [2.0.0] - 2022-05-05
+
 ### Changed
+
 - [General] Updated dev dependencies.
 - [General] React 18 peer dependency.
 - [General] Recoil 0.7 peer dependency.
 
 ## [1.0.1] - 2021-09-01
+
 ### Added
+
 - [List] Added `UseRecoilListResult` type describing the result of `useRecoilList` hook.
 - [Dialog] Added `UseRecoilDialogResult` type describing the result of `useRecoilDialog` hook.
 - [Filters] Added `UseRecoilFiltersResult` type describing the result of `useRecoilFilters` hook.